+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Release date: ??.??.????
 * Updated styleguide (1.7.0 -> 1.13.0) (HUB-412)
 * Degree programme specific news feed (HUB-378)
 * Change news archive title from News to Current topics (HUB-390)
+* Removed "For new students" checkbox from instructions (HUB-426)
 
 
 ## 1.29